refactor: organize Playwright tests by auth state
Split tests into auth, migration, and session files to improve clarity and avoid storageState conflicts. No behavior changes.
